refactor: components 분리 및 정리 (#36)
