+++ frontend/src/components/home/HomeHotReviews.vue
@@ -1,0 +1,22 @@
+<template>
+  <div class="hot-reviews">
+    <h3>Hot Reviews</h3>
+    <div></div>
+  </div>
+</template>
+
+<script setup>
+import { onMounted } from 'vue'
+
+// hot review 불러오기
+const hot_review = function() {
+
+}
+
+onMounted(() => {
+  hot_review()
+})
+</script>
+
+<style scoped>
+</style>
