+++ frontend/src/components/movies/MovieListFilter.vue
@@ -1,0 +1,16 @@
+<template>
+  <div class="filter-container">
+    <div class="dropdown">장르</div>
+    <div>연도</div>
+    <hr>
+  </div>
+</template>
+
+<script setup>
+</script>
+
+<style scoped>
+.filter-container {
+  margin-bottom: 20px;
+}
+</style>
